requirements.txt
anthropic>=0.40.0
openai>=1.50.0
yfinance>=0.2.40
requests>=2.32.0
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
